Push the sugar into the parent classes.
This implies making all of DB, Cursor and Row inherit from a new base
class, HasDB. Properly speaking this ought to be a role rather than a
base class, but I'm not sure I want to convert all this to Moose just
yet.

A statement like

    use DBIx::OQM "DB";

will now import all the sugar from %SUGAR in every parent class, and
arrange to unimport it at the end of the scope.

As a side-effect of moving the sugar out of DBIx::OQM, the %P hash has
been replaced with a proper pair of register/lookup functions.