I forgot to name the subs.
I'm increasingly dubious about Sub::Name. Quite apart from the nasty
hacks to work around the fact it turns CvANON off (and at least ~ magic
is a *little* better than stuffing something into the @_ slot in the
pad), it will rename a CV into a different package. It isn't clear (to
me) what effect this has, in particular which builtins will see the old
package and which the new.