Separate 'cursor' from 'query'.
'cursor' returns a Cursor with which to read the results, 'query'
returns all the results immediately. (Of course, currently the Cursor
doesn't actually use a cursor (well, it probably uses the Pg unnamed
portal underneath somewhere...), but that's the next step.)

As part of this, move the Defer logic and the magic variables into a new
Query class. Since (unlike String::Defer) I don't have to worry about an
argumentless stringify method, I can pass the data needed for expansion
properly rather than using a global.