Real cursors (at least for Pg).
This needs a small amount of db-specific state, so DB now holds a driver
object which handles interaction with the DBI. Possibly the
DBIx::Connector object should be subsumed into that.

Databases without a specific driver get a fallback implementation that
fetches all the rows in one go and fakes the cursor client-side. I
suppose I could use DBI statement handles, on the assumption that they
might implement cursors underneath; perhaps a ::DBI superclass for
drivers that do work like that?