Allow Row classes to extend other Row classes.
I'm not completely happy with the way this works. The class is set up to
inherit from Row by DBIx::OQM->import, and then the 'extends' sugar has
to remove that superclass and replace it with the new list. There is
also no support for inheriting from non-DBIx::OQM superclasses, which
would probably be a useful feature. (Though, at some point I think I
start saying 'this is a has-a, not an is-a, relationship'.)