setup_row_class sugar for Rows with no methods.
A lot of Row classes end up with no more than a 'columns' declaration.
Add sugar to HasDB to generate a class like that, to avoid the need for
hundreds of tiny files.

This is implemented with an @INC hook, since that seems to be the
easiest way to get everything to agree this should be treated as a real
class. I'm not certain it belongs in HasDB: it's only there since it's
the parent of both Row and DB. I think perhaps it should be directly in
::OQM, or perhaps in a Util module that just exports sugar.