'detail' and 'action' query sugars.
detail expects a query which returns a single column, and returns that
value without wrapping it in an object. If called in scalar context it
returns only the first row, and warns if there are more that have been
ignored.

action executes a piece of SQL using 'do'. It's necessary for DML like
INSERT/UPDATE/DELETE (without RETURNING: INSERT RETURNING can just be an
ordinary 'query' or 'cursor' as far as we're concerned).

The implementation here is a bit messy: there's too much duplication.
Tidy up later...