Allow QuerySets.
QuerySet is just HasDB renamed, but now it is a concrete class.
(QuerySet is a more useful (outward-looking) name anyway.) A QuerySet
provides a way to avoid stuffing all the top-level entry point queries
into the DB, instead grouping them in sensible groups.

I would prefer to be able to make top-level queries class methods on
their Row classes, since they are effectively constructors, but I don't
think that's possible given the need to get hold of a DB handle.