Inflators.
I'm not sure this is the best implementation. While classes seem nice,
it's a lot of overhead for what is (in the end) a single function, and
it makes the question of loading app-specific inflators a little tricky.
(What are they called? How do we find them? Having to say "+Prospect::
DB::Inflate::PartialDate" all the time would be infuriating, but OTOH
I've so far avoided making assumptions about the user's package
structure beyond everything being under the DB.) It's possible it would
be better just to maintain a hash of subrefs in Inflate.pm, with a
number of inflators pre-registered.

Currently only inflators are supported, not deflators: I'm not sure what
the best way to support deflators would be. Obviously they would need to
be invoked in $Self and $Arg, but it's not clear to me how to detect the
things that need deflating.

One (obvious) solution would be to only support them in $Self, and only
for columns registered as having an inflator. This is so limited I
suspect it's useless: it means, among other things, that you can't pass
a value from one Row to a method on another without deflating it
manually. The inconsistency would be infuriating.

The ideal solution would look at all values returned from placeholder{},
and if they were deflatable objects, deflate them. The tricky part is
identifying deflatable objects, and matching them to deflators. I
believe the only correct answer would be to check every object against
every registered deflator with ->isa, which would be horrendously slow.
Even that doesn't tell us what to do if an object ->isa two classes with
different deflators.

One possible answer would be, *if* an object uses UNIVERSAL::isa (which
will be usual), to get the linearized ISA and check it against our
registrations. The fallback case will still be slow, but uncommon. Of
course, this is still a whole lot of checking for objects which don't
deflate at all.

Another, of course, would be to wrap inflated objects in a proxy which
knows how to deflate them again. This would simplify things enormously,
at the cost of not allowing the user to pass objects in from the
outside. YA would be to require queries to specify their deflated
arguments explicitly, though I suspect that would get boring very fast.