Use subrefs as inflators, not classes.
It simplifies just about everything, and most inflators are going to be
so simple it's not worth doing anything else.

It does occur to me that e.g. DateTime will need a little more than
that. I think the right plan here is: if register_inflators is passed a
string instead of a coderef, that string is a classname. If any column
in any Row uses that inflator, that fact is registered in the DB. When
the DB connects (and, I think, probably every time it reconnects), it
goes through its list of inflator classes and instantiates them all,
passing itself so the inflator class can talk to the db if it needs to.