Expand Querys recursively.
That is, if a component of an expanded Query is itself another Query,
concatenate it in and expand again. It's a little tricky to get the list
of bind values at the right moment, and I should probably put in some
sort of recursion limit, but it now allows e.g. $Arg[0] to expand to a
Query, which will then be expanded again.

I really need to take the time to tidy Query up a bit, it's getting out
of control. I think, for a start, I need a Query::Expander class, which
handles the expansion.