Row::Generic, for when we don't know the columns.
This implements a generic Row class, that gets its column names from the
DBI and implements them with AUTOLOAD. It also extends the interface of
Row->_new to be

    Row->_new($DB, \@row, [\@cols])

Currently it won't work with Cursors because Cursor doesn't know how to
fetch the column names. In order to do that I'll need to find some way
to get hold of the sth used by Driver->fetch; either that, or
fetch-and-discard a row in Driver->cursor, just to get the columns.

Implementing AUTOLOAD/can correctly seems to be surprisingly difficult.
Perhaps I should write a general solution...