Make the QS sugar more sugary.
The bulk of the logic has moved into ->do_* methods on DB, with the QS
sugar just installing little wrappers round that. This is not just
cleaner, it's also directly useful: when building queries dynamically
there are occasions when the old system would require something like

    query do_sql => "" => $Arg[0];

which is just silly.