Add [@%]ArgQ and %SelfQ.
These are identifier-quoted versions of ArgX and SelfX.

Also rationalise Arg{,X,Q} and Self{,X,Q}, using %P and %Q. This means
more qexs in appropriate places. I wonder if, if I were to make defer
and placeholder call qex, I could get rid of the recursive expansion in
expand? I'd be replacing it with Perl-level recursion, of course, but I
don't mind that.