Allow Query->expand to redefer.
This makes it possible to partially expand a Query with the currently-
available information, then finish later. In particular, this will be
needed to allow class methods which return Queries.

Currently this is used by %Q and %P; both redefer if the hash passed to
->expand didn't include a DB. This means that contexts which want a full
expansion will now need to explicitly pass in a 'db' parameter. Change
QS to do this.