'method' sugar on QuerySet.
This installs either a sub or a Query as a method. Installing a plain
sub like this has the advantage that it won't conflict with sugar subs
that are going to disappear before runtime.

Installing a Query gives you a method which expands that Query, but
without passing a 'db' parameter. This means it will effectively close
over the current %Self and @Args, leaving %P and %Q to expand later when
we've got a DB object to run them against.