Replace Exporter with exports from Irian.
This means I can clean up properly afterwards. It also means I get an
error if two modules try to export the same function, which seems like a
good idea.

Currently Query still uses Exporter for the variables, mostly because I
can't be bothered to replace it. I suspect a replacement based on
find_sym would be rather easy, and it would let me make the variable
exports lexically-scoped as well.

(I am assuming here that the replace-the-glob trick works for variables
as well as for functions. I see no reason it shouldn't: the glob is
compiled into the optree in exactly the same way. *However*, I do need
to check that is still the case under ithreads.)