Move the bulk of t/basic.t into a module.
These are all generic QS tests, so I will want to run them against other
classes as well.

Factoring out the sugar calls was a little tricky, since they have to be
within the lexical scope of the 'use DBIx::Irian'. I ended up with a
template-expanding @INC hook, which is a little more complicated that I
would prefer for a test, but I don't really see any other option.

In Real Life, I believe it should be possible to 'use DBIx::Irian
"QuerySet"' in a required file and, since the package in question should
already inherit from QS, this should do nothing but export the sugar. I
should probably test that.