Make expand_query look up Rows for itself.
Previously, it was necessary for a caller of expand_query to call
lookup() on the row class. Since lookup is a private interface, that
meant expand_query would have had to be considered private as well.

Fix this by making expand_query take a class name instead, and do its
own lookups.