Test the do_* methods on the DB.
The easiest way to do this is to reuse the QS tests, which means fudging
the loading of t/QS.pl so a statement like

    query foo => ...;

will be recorded in a hash instead of installing a method. This is a bit
more tricky than I really like in a test; it might be better to make
t/QS.pl hold a data structure in the first place and then generate the
code from there.