Support retrying serialisable transactions.
This requires two levels of retry logic: one inside DBC, to retry if the
connection drops, and one here to retry serialisation failures. It also
needs a new 'txnmode' config parameter, specifying isolation levels and
so on.

Obviously most of the actual work has to happen in the driver. I've
tried to at least gesture in the direction of making this generic: the
Driver base class supports read-only txns by setting the ->{ReadOnly}
DBI attribute; this means we need a mechanism to reset any such settings
at the end of the txn. (Pg forgoes this in favour of issuing SET
TRANSACTION directly, since the latter will be rolled back by the
database.) The driver is allowed to retry given any error it thinks it
can recover from; currently the only error Pg will retry is
serialisation failure.

Retrieving the SQLSTATE is also a little tricky. The only right thing to
to is to use something like Exception::Class::DBI, to catch the DBI
variables at the point where the error is thrown: otherwise if anything
in the stack-unwind makes any DBI calls the information in the dbh will
be invalid. However, I don't want to force people to use ECDBI, not
least because it's kinda heavyweight, so I provide three alternatives:
use plain RaiseError, and pray the handle status is still right; use a
HandleError which throws a specified ECDBI-compatible exception class;
or supply an arbitrary coderef to extract the DBI statusses from the dbh
and the exception.