Add a Template::Iterator for Cursor.
Cursor::TT2 is a subclass of Template::Iterator which iterates over a
Cursor. This lets me do [% FOR r IN cursor %] in a template without
having to prefetch all the rows.

This implementation is rather minimal. I suspect I will never be able to
get ->size to work, given that even the DB doesn't know how many rows it
will be returning until it's returned them all, but the other methods
could be implemented with some client-side counters. I'm not sure
whether they belong here or in Cursor. (Probably: count and last belong
in Cursor, assuming last can be implemented at all; possibly also prev,
though this will make the buffering a bit more complicated. The rest
belong here.)

(I suppose I could implement size using a scrollable cursor and MOVE
ALL; MOVE FIRST; to count the rows. This would probably defeat some of
the point of using a cursor in the first place, but at least the
calculated rows would stay in the database rather than being sent over.)